# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Runtime API
  version: 0.3.0
servers:
  - url: /v1
security:
  - serviceToken: []
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
  schemas:
    TargetRef:
      type: object
      required: [type, id]
      properties:
        type: { type: string }
        id: { type: string }
        display:
          type: object
          properties:
            title: { type: string }
            url: { type: string }
        metadata:
          type: object
          additionalProperties: true
    WorkspaceLease:
      type: object
      required: [id, root_path]
      properties:
        id: { type: string }
        provider: { type: string }
        root_path: { type: string }
        cleanup_policy: { type: string, enum: [always, on_terminal, manual] }
        metadata:
          type: object
          additionalProperties: true
    SkillRef:
      type: object
      properties:
        skill_id: { type: string }
        key: { type: string }
        version: { type: string }
        version_key: { type: string }
        config:
          type: object
          additionalProperties: true
    Agent:
      type: object
      required: [app_id, name]
      properties:
        id: { type: string }
        app_id: { type: string }
        name: { type: string }
        runtime_kind: { type: string, description: "Use native_sdk for writes. Historical records may contain codex or opencode." }
        provider: { type: string }
        model: { type: string }
        system_prompt: { type: string }
        skills:
          type: array
          items: { $ref: '#/components/schemas/SkillRef' }
        allowed_tools:
          type: array
          items: { type: string }
        allowed_targets:
          type: array
          items: { type: string }
        approval_mode: { type: string, enum: [never, risk_based, mutating_tools, always] }
        default_invocation_mode: { type: string, enum: [autonomous, interactive] }
        execution_config:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    StartRunRequest:
      type: object
      required: [app_id, agent_id, target]
      additionalProperties: false
      properties:
        app_id: { type: string }
        host_run_id: { type: string }
        agent_id: { type: string }
        target: { $ref: '#/components/schemas/TargetRef' }
        instructions: { type: string }
        allowed_tools:
          type: array
          items: { type: string }
        external_actor_id: { type: string }
        mode: { type: string, enum: [autonomous, interactive] }
        execution_mode: { type: string, enum: [lightweight, durable] }
        trigger:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        turn_policy:
          $ref: '#/components/schemas/TurnPolicy'
        mcp_servers:
          type: array
          maxItems: 16
          items: { $ref: '#/components/schemas/RunMCPServer' }
    RunMCPServer:
      type: object
      required: [server_id, server_name, transport, url, tools]
      additionalProperties: false
      properties:
        server_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9][A-Za-z0-9_.:-]*$'
        server_name:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
        transport:
          type: string
          enum: [streamable_http]
        url:
          type: string
          format: uri
          pattern: '^https://'
        tools:
          type: array
          minItems: 1
          maxItems: 128
          items: { $ref: '#/components/schemas/RunMCPTool' }
        skills:
          type: array
          maxItems: 16
          items: { $ref: '#/components/schemas/SkillRef' }
        credential:
          $ref: '#/components/schemas/RunMCPCredential'
    RunMCPTool:
      type: object
      required: [name, access]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9][A-Za-z0-9_.-]*$'
        access:
          type: string
          enum: [read, write]
    RunMCPCredential:
      type: object
      required: [type]
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [bearer_token, headers]
        access_token:
          type: string
          minLength: 1
          writeOnly: true
        headers:
          type: object
          minProperties: 1
          additionalProperties:
            type: string
            minLength: 1
          writeOnly: true
        expires_at:
          type: string
          format: date-time
      oneOf:
        - required: [access_token]
          properties:
            type: { const: bearer_token }
        - required: [headers]
          properties:
            type: { const: headers }
...
    AgentRun:
      type: object
      required: [id, app_id, agent_id, target, runtime_kind, execution_mode, invocation_mode, status]
      properties:
        id: { type: string }
        app_id: { type: string }
        host_run_id: { type: string }
        agent_id: { type: string }
        target: { $ref: '#/components/schemas/TargetRef' }
        runtime_kind: { type: string }
        execution_mode: { type: string }
        invocation_mode: { type: string }
        external_actor_id: { type: string }
        status: { type: string, enum: [queued, running, paused, completed, failed, cancelled] }
        pause_reason: { type: string }
        approval_state: { type: string }
        input:
          type: object
          additionalProperties: true
        output_summary:
          type: object
          additionalProperties: true
        workspace_lease:
          $ref: '#/components/schemas/WorkspaceLease'
        error_message: { type: string }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    TurnPolicy:
      type: object
      additionalProperties: false
      properties:
        mode: { type: string, enum: [complete_on_finish, pause_after_assistant] }
        idle_timeout_seconds: { type: integer, minimum: 0 }
        expired_resume_strategy: { type: string }
        completion_mode: { type: string, enum: [implicit, explicit_finish] }
        max_completion_corrections: { type: integer, minimum: 0, maximum: 5 }
...
    RunPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AgentRun' }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
...
paths:
...
  /agents:
    get:
      parameters:
        - name: app_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Agent' }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Agent' }
      responses:
        '201':
          description: Created agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
  /agents/{agent_id}:
    get:
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
        - name: app_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
    put:
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
        - name: app_id
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Agent' }
      responses:
        '200':
          description: Updated agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '201':
          description: Created agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
  /runs:
    get:
      parameters:
        - name: app_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AgentRun' }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartRunRequest' }
      responses:
        '202':
          description: Started run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRun' }
  /runs/search:
    get:
      parameters:
        - name: app_id
          in: query
          required: true
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 25, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Filtered run page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunPage' }
  /runs/{run_id}:
    get:
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
        - name: app_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRun' }
...